refactor(backend): extract tag crud module
Move the seven tag functions out of crud/__init__.py into crud/tag.py,
following the split already done for comment, token, metric and others.
Consumers import the module directly; nothing is re-exported.

test_service_security.py reached these through hasattr/getattr(crud,
name), which would have silently stopped asserting once they moved, so
its function list now carries the owning module alongside each name.
