+++ tests/backend/crud/test_transaction_management.py
@@ -32,6 +32,7 @@ from sqlalchemy.orm import Session
 
 from rhesis.backend.app import crud, models, schemas
 from rhesis.backend.app.constants import EntityType
+from rhesis.backend.app.crud import tag as tag_crud
 from rhesis.backend.app.crud.comment import add_emoji_reaction, remove_emoji_reaction
 from tests.backend.routes.fixtures.data_factories import (
     CommentDataFactory,
@@ -160,7 +161,7 @@ class TestCRUDTransactionManagement:
         entity_id = behavior.id
 
         # Assign tag
-        result = crud.assign_tag(test_db, tag_create, entity_id, entity_type, test_org_id)
+        result = tag_crud.assign_tag(test_db, tag_create, entity_id, entity_type, test_org_id)
 
         # Verify tag was created and persisted
         assert result is not None
@@ -203,7 +204,7 @@ class TestCRUDTransactionManagement:
         entity_type = EntityType.BEHAVIOR
 
         # Assign tag
-        created_tag = crud.assign_tag(test_db, tag_create, entity_id, entity_type, test_org_id)
+        created_tag = tag_crud.assign_tag(test_db, tag_create, entity_id, entity_type, test_org_id)
         tag_id = created_tag.id
 
         # Verify tagged item exists
@@ -215,7 +216,7 @@ class TestCRUDTransactionManagement:
         assert initial_tagged_items == 1
 
         # Remove tag
-        result = crud.remove_tag(test_db, tag_id, entity_id, entity_type, test_org_id)
+        result = tag_crud.remove_tag(test_db, tag_id, entity_id, entity_type, test_org_id)
 
         # Verify tag was removed (committed)
         assert result is True
