+++ tests/backend/utils/test_query_builder_load.py
@@ -16,6 +16,7 @@ from sqlalchemy import inspect as sa_inspect
 from sqlalchemy.orm import Session
 
 from rhesis.backend.app import crud, models, schemas
+from rhesis.backend.app.crud import tag as tag_crud
 from rhesis.backend.app.crud.metric import get_metrics
 from rhesis.backend.app.utils import crud_utils
 from rhesis.backend.app.utils.query_utils import QueryBuilder, include
@@ -189,7 +190,7 @@ class TestMetricBehaviorNestedM2MLoads:
                 user_id=authenticated_user_id,
             )
         )
-        crud.assign_tag(
+        tag_crud.assign_tag(
             db=test_db,
             tag=schemas.TagCreate(name="nested-behavior-tag"),
             entity_id=behavior.id,
@@ -242,7 +243,7 @@ class TestMetricBehaviorNestedM2MLoads:
                 user_id=authenticated_user_id,
             )
         )
-        crud.assign_tag(
+        tag_crud.assign_tag(
             db=test_db,
             tag=schemas.TagCreate(name="nested-metric-tag"),
             entity_id=metric.id,
